command/services: move the config helpers to parent package
