+++ command/services/config_test.go
@@ -1,4 +1,4 @@
-package register
+package services
 
 import (
 	"testing"
